pyproject.toml
[project]
name = "outbound-caller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-cartesia",
    "livekit-plugins-deepgram~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai~=1.2",
    "python-dotenv",
    "pymongo[srv]~=4.10",
]
